Move calculator to flow directory - calculate-agricultural-holiday-entitlement
This calculator is specific to the calculate-agricultural-holiday-entitlement
flow and so should be co-located.

* SmartAnswer::Calculators:AgriculturalHolidayEntitlementCalculator
  -> SmartAnswer::CalculateAgriculturalHolidayEntitlementFlow::Calculator

* SmartAnswer::Calculators::AgriculturalHolidayEntitlementCalculatorTest
  -> SmartAnswer::CalculateAgriculturalHolidayEntitlementFlow::CalculatorTest

For some reason the calculator was not in the regression test checksums file, so
I've added it in the new location.
