+++ lib/smart_answer_flows/calculate-agricultural-holiday-entitlement.rb
@@ -1,3 +1,5 @@
+require 'smart_answer_flows/calculate-agricultural-holiday-entitlement/calculator'
+
 module SmartAnswer
   class CalculateAgriculturalHolidayEntitlementFlow < Flow
     def define
@@ -6,7 +8,7 @@ module SmartAnswer
       status :published
       satisfies_need "100143"
 
-      calculator = Calculators::AgriculturalHolidayEntitlementCalculator.new
+      calculator = Calculator.new
 
       multiple_choice :work_the_same_number_of_days_each_week? do
         option "same-number-of-days"
